#3 dissociate userDetailsService from crud UserService
