+++ back/src/main/java/org/hakkou/mock/boats/service/UserService.java
@@ -1,20 +1,12 @@
 package org.hakkou.mock.boats.service;
 
 import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
-
 import org.hakkou.mock.boats.dto.UserDto;
 import org.hakkou.mock.boats.exceptions.UserException;
 import org.hakkou.mock.boats.mappers.UserMapper;
 import org.hakkou.mock.boats.model.User;
 import org.hakkou.mock.boats.repo.UserRepository;
 import org.hakkou.mock.boats.service.management.UserManagement;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,7 +16,7 @@ import lombok.AllArgsConstructor;
 @Service
 @Transactional
 @AllArgsConstructor
-public class UserService implements UserManagement, UserDetailsService {
+public class UserService implements UserManagement {
 
     private final UserRepository userRepository;
 
@@ -77,25 +69,5 @@ public class UserService implements UserManagement, UserDetailsService {
         String encodedPassword = passwordEncoder.encode(inputUser.getPassword());
         inputUser.setPassword(encodedPassword);
         return inputUser;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {        
-        Optional<User> repoUser = userRepository.findByUsername(username);
-        if(repoUser.isPresent()) {
-            return userToUserDetails(repoUser.get());
-        } else {
-            throw new UsernameNotFoundException("user with username : " + username + " not found");
-        }
-    }
-
-    private org.springframework.security.core.userdetails.User userToUserDetails(User entityUser) {
-        Set<GrantedAuthority> authorities = entityUser.getRoles().stream()
-        .map(role -> new SimpleGrantedAuthority(role.getName()))
-        .collect(Collectors.toSet());
-
-        return new org.springframework.security.core.userdetails.User(
-            entityUser.getUsername(), entityUser.getPassword(), authorities
-        );
-    }    
+    }        
 }
