+++ back/src/main/resources/application.properties
@@ -12,7 +12,3 @@ spring.jpa.hibernate.ddl-auto= update
 spring.h2.console.enabled=true
 # default path: h2-console
 spring.h2.console.path=/h2-ui
-
-spring.security.user.name=admin
-spring.security.user.password=admin
-
